requirements.txt
alembic==1.17.1
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
ecdsa==0.19.1
fastapi==0.120.2
greenlet==3.2.4
h11==0.16.0
httptools==0.7.1
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
pillow==12.0.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyMySQL==1.1.2
PyPDF2==3.0.1
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
reportlab==4.4.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
email-validator==2.2.0
passlib[bcrypt]==1.7.4